# docs/api/comments.openapi.yaml
openapi: 3.0.3
info:
  title: Comments API
  version: '1.0.0'
  description: 用户对动物评论 + AI 后台处理
paths:
  /v1/comments:
    post:
      summary: 发表评论
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCommentDto'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Comment'
        '400': { description: Bad Request }
  /v1/comments/animal/{animal_id}/summary:
    get:
      summary: 评论摘要(情感分布 + Top 关键词)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: animal_id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommentSummary'
  /v1/comments/animal/{animal_id}:
    get:
      summary: 列出该动物评论
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: animal_id
          required: true
          schema: { type: string }
        - in: query
          name: limit
          schema: { type: integer, default: 20, maximum: 100 }
        - in: query
          name: offset
          schema: { type: integer, default: 0 }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  total: { type: integer }
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/Comment' }
  # ========== P3 (2026-07-07) ==========
  # admin clue review endpoints (extension, original /v1/comments* unchanged)
  /v1/admin/clues/pending:
    get:
      summary: list pending clues
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClueListResponse'
  /v1/admin/clues/{animal_id}/{match_id}/decide:
    post:
      summary: decide a clue (confirmed/rejected)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: animal_id
          required: true
          schema: { type: string }
        - in: path
          name: match_id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [decision]
              properties:
                decision:
                  type: string
                  enum: [confirmed, rejected]
                note: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  match_id: { type: string }
                  animal_id: { type: string }
                  decision:
                    type: string
                    enum: [confirmed, rejected]

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Sentiment:
      type: string
      enum: [care, seek, fake, thanks, report, neutral]
    Comment:
      type: object
      properties:
        comment_id: { type: string, format: uuid }
        animal_id: { type: string, format: uuid }
        reporter_id: { type: string, format: uuid }
        content: { type: string, maxLength: 500 }
        sentiment: { $ref: '#/components/schemas/Sentiment' }
        is_hidden: { type: boolean }
        created_at: { type: string, format: date-time }
    CreateCommentDto:
      type: object
      required: [animal_id, content]
      properties:
        animal_id: { type: string, format: uuid }
        content: { type: string, minLength: 1, maxLength: 500 }
    CommentSummary:
      type: object
      properties:
        animal_id: { type: string, format: uuid }
        total: { type: integer }
        sentiment_dist:
          type: object
          additionalProperties: { type: integer }
        top_keywords:
          type: array
          items: { type: string }
        auto_summary: { type: string }

    ClueListResponse:
      type: object
      properties:
        total: { type: integer }
        pending_count: { type: integer }
        confirmed_count: { type: integer }
        rejected_count: { type: integer }
        items:
          type: array
          items: { $ref: '#/components/schemas/ClueItem' }
    ClueItem:
      type: object
      properties:
        match_id: { type: string }
        comment_id: { type: string }
        animal_id: { type: string }
        comment_reporter_id: { type: string }
        sentiment: { $ref: '#/components/schemas/Sentiment' }
        keywords:
          type: array
          items: { type: string }
        created_at: { type: string, format: date-time }
        candidate_event_id: { type: string }
        candidate_event_reporter_id: { type: string }
        candidate_event_address: { type: string }
        match_score: { type: number }
        match_reasons:
          type: array
          items: { type: string }
        status:
          type: string
          enum: [pending, confirmed, rejected]
        recorded_at: { type: string, format: date-time }
        decided_by: { type: string }
        decided_at: { type: string, format: date-time }
        decision_note: { type: string }